migrate home feature
gets renamed to app feature
